MyIA.AI.Notebooks/GenAI/Vibe-Coding/Roo-Code/Corrections/corrigés ateliers-roo/Ateliers avancés/solution-matching-cv/requirements.txt
# Fichier de dépendances Python pour le projet solution-matching-cv

# --- Application ---
flask
python-dotenv
semantic-kernel
pandas
numpy
scikit-learn
Werkzeug
matching
chromadb

# --- Tests ---
pytest
playwright
pytest-playwright
pytest-html